review response

+++ src/main/java/com/worksap/stm2016/service/user/UserProfileService.java
@@ -1,27 +1,12 @@
 package com.worksap.stm2016.service.user;
 
-import com.worksap.stm2016.domain.User;
-import com.worksap.stm2016.domain.UserProfile;
-import com.worksap.stm2016.enums.Role;
-import com.worksap.stm2016.enums.UserStatus;
-import com.worksap.stm2016.form.UserCreateForm;
-import com.worksap.stm2016.form.UserRegisterForm;
-import com.worksap.stm2016.repository.UserProfileRepository;
-import com.worksap.stm2016.repository.UserRepository;
-import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
-import org.json.simple.parser.ParseException;
+import com.worksap.stm2016.domain.user.User;
+import com.worksap.stm2016.domain.user.UserProfile;
+import com.worksap.stm2016.repository.user.UserProfileRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.Iterator;
-
-import static com.worksap.stm2016.specification.BasicSpecs.*;
 
 @Service
 public class UserProfileService {
@@ -29,24 +14,25 @@ public class UserProfileService {
     private static final Logger LOGGER = LoggerFactory.getLogger(UserProfileService.class);
 
     @Autowired
-    private UserService userService;
-    @Autowired
     private UserProfileRepository userProfileRepository;
 
     public UserProfile get(Long id){
-        UserProfile userProfile = userProfileRepository.findOne(id);
-        userProfile.setName(userService.get(id).getName());
-        return userProfile;
-    }
-    public void saveOrUpdate(UserProfile userProfile){
-        userProfileRepository.save(userProfile);
-        User user = userService.get(userProfile.getId());
-        user.setName(userProfile.getName());
-        userService.update(user);
+        return userProfileRepository.findOne(id);
     }
 
-    void delete(Long id){
-        userProfileRepository.delete(id);
+    public UserProfile update(UserProfile userProfile){
+        return userProfileRepository.save(userProfile);
     }
 
+    public Long delete(Long id){
+        if (userProfileRepository.findOne(id) == null) {
+            return Long.valueOf(-1); // Not exists
+        }
+        try {
+            userProfileRepository.delete(id);
+        } catch (Exception e) {
+            return id; // Referenced
+        }
+        return Long.valueOf(0); // OK
+    }
 }
